+++ shelkovnikov.dmitrii/I3/removeextraspaces.h
@@ -1,4 +1,4 @@
 #ifndef SPBSPU_LABS_2022_AAP_904_A_REMOVEEXTRASPACES_H
 #define SPBSPU_LABS_2022_AAP_904_A_REMOVEEXTRASPACES_H
-char *remove_extra_spaces(char *destination, const char *source, int *res_size);
+char *remove_extra_spaces(char *destination, const char *source, size_t *res_size);
 #endif
